recalculate deltas to fix upstream bug
When a day is missing, the delta is calculated from 0 by upstream.
I fix this by recalculating the sums based on the n'th data value
in the data set.  THIS STILL IS NOT CORRECT, because if one day
is missing in a week, we calculate the delta based on 8 days
in this case.  But this is better than nothing.

The real fix would be to introduce 'dummy days' which just
interpolate missing days.  But this requires knowledge of then
data structures.  Currently this viewer here is structure
very agnostic to what upstream data it has to show.